Fixed SQLite extension bug
There was an issue when calling the SqliteConnection class and setting the file name with './' (Example: SqliteConnection("connection", "./test.db")), where only ".db" was generated instead of "test.db". Upon analyzing the code, I noticed that there is a split to obtain the file extension and check if it exists. However, this split returns 3 elements instead of 2, with the first element being empty. Additionally, this could break if someone includes ".sqlite", for example.

My solution is, instead of splitting the file, to simply grab the last 3 characters (.db) and use the "File" object itself within the setNameFile function to split the extension without compatibility issues.

Also in this update, a new parameter was added within the ConnectionManager called disableLogs.

As the name suggests, the option toggles the logging generated by the API on or off. Why did I do this? Well, I started to think and this could bring a problem of high disk consumption for those with few resources. However, it is good because it saves a guarantee of data that that Transaction was executed, so simply not to remove it from the API, a new parameter allowing the developer to turn it on or off was added, the default value is "false", meaning it will come with the logs active.